changed z/x/y from uint64 to uint

+++ atlas/atlas.go
@@ -46,7 +46,7 @@ func (a *Atlas) AllMaps() []Map {
 
 //	SeedMapTile will generate a tile and persist it to the
 //	configured cache backend
-func (a *Atlas) SeedMapTile(ctx context.Context, m Map, z, x, y uint64) error {
+func (a *Atlas) SeedMapTile(ctx context.Context, m Map, z, x, y uint) error {
 	//	confirm we have a cache backend
 	if a.cacher == nil {
 		return ErrMissingCache
@@ -63,9 +63,9 @@ func (a *Atlas) SeedMapTile(ctx context.Context, m Map, z, x, y uint64) error {
 	//	cache key
 	key := cache.Key{
 		MapName: m.Name,
-		Z:       uint64(z),
-		X:       uint64(x),
-		Y:       uint64(y),
+		Z:       z,
+		X:       x,
+		Y:       y,
 	}
 
 	return a.cacher.Set(&key, b)
@@ -80,9 +80,9 @@ func (a *Atlas) PurgeMapTile(m Map, tile *tegola.Tile) error {
 	//	cache key
 	key := cache.Key{
 		MapName: m.Name,
-		Z:       uint64(tile.Z),
-		X:       uint64(tile.X),
-		Y:       uint64(tile.Y),
+		Z:       tile.Z,
+		X:       tile.X,
+		Y:       tile.Y,
 	}
 
 	return a.cacher.Purge(&key)
@@ -157,7 +157,7 @@ func SetCache(c cache.Interface) {
 
 //	SeedMapTile will generate a tile and persist it to the
 //	configured cache backend for the DefaultAtlas
-func SeedMapTile(ctx context.Context, m Map, z, x, y uint64) error {
+func SeedMapTile(ctx context.Context, m Map, z, x, y uint) error {
 	return DefaultAtlas.SeedMapTile(ctx, m, z, x, y)
 }
